一行ずつ表示の隙間が空くバグを修正

+++ src/fonts.rs
@@ -9,20 +9,20 @@ unsafe fn get_font(character: u8) -> &'static [u8; 16usize] {
     &fonts[char_idx]
 }
 
-pub fn write_byte(x: u32, y: u32, character: u8, writer: &PixelWriter) {
+pub fn write_byte(x: usize, y: usize, character: u8, writer: &PixelWriter) {
     let font_slice = unsafe { get_font(character) };
     for dy in 0..16 {
         for dx in 0..8 {
             if ((font_slice[dy] << dx) & 0b10000000) == 0b10000000 {
-                writer.write(x + dx as u32, y + dy as u32, &PixelColor::White);
+                writer.write(x + dx, y + dy, &PixelColor::White);
             } else {
-                writer.write(x + dx as u32, y + dy as u32, &PixelColor::Black);
+                writer.write(x + dx, y + dy, &PixelColor::Black);
             }
         }
     }
 }
 
-pub fn write_string(x: u32, y: u32, s: &str, writer: &PixelWriter) {
+pub fn write_string(x: usize, y: usize, s: &str, writer: &PixelWriter) {
     let mut dx = 0;
     for i in 0..s.len() {
         write_byte(x + dx, y, s.as_bytes()[i], writer);
